# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Gfolio Admin & Corporate Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /gifts:
    get:
      operationId: listGifts
      tags: [gifts]
      summary: List gift orders
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, delivered, redeemed, expired, cancelled, all]
        - name: type
          in: query
          schema:
            type: string
            enum: [digital, physical, all]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Gifts list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GiftListResponse"
  /gifts/{giftId}:
    get:
      operationId: getGiftById
      tags: [gifts]
      summary: Get gift details
      parameters:
        - name: giftId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Gift details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GiftOrder"
  /gifts/{giftId}/resend:
    post:
      operationId: resendGift
      tags: [gifts]
      summary: Resend a gift
      parameters:
        - name: giftId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Gift resent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GiftOrder"
  /gifts/{giftId}/cancel:
    post:
      operationId: cancelGift
      tags: [gifts]
      summary: Cancel a gift
      parameters:
        - name: giftId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Gift cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GiftOrder"
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List transactions with filters
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [success, failed, pending, refunded, all]
        - name: type
          in: query
          schema:
            type: string
            enum: [purchase, gift, refund, settlement, all]
        - name: search
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Transactions list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionListResponse"
  /transactions/{txnId}:
    get:
      operationId: getTransactionById
      tags: [transactions]
      summary: Get transaction details
      parameters:
        - name: txnId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Transaction details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{txnId}/retry:
    post:
      operationId: retryTransaction
      tags: [transactions]
      summary: Retry a failed transaction
      parameters:
        - name: txnId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Transaction retried
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{txnId}/refund:
    post:
      operationId: refundTransaction
      tags: [transactions]
      summary: Initiate a refund
      parameters:
        - name: txnId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Refund initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List corporate campaigns
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, scheduled, active, completed, cancelled, all]
      responses:
        "200":
          description: Campaigns list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignListResponse"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a new campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignRequest"
      responses:
        "201":
          description: Campaign created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{campaignId}:
    get:
      operationId: getCampaignById
      tags: [campaigns]
      summary: Get campaign details
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Campaign details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
...
components:
  schemas:
...
    GiftListResponse:
      type: object
      properties:
        gifts:
          type: array
          items:
            $ref: "#/components/schemas/GiftOrder"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - gifts
        - total
        - page
        - limit
    GiftOrder:
      type: object
      properties:
        id:
          type: string
        senderId:
          type: string
        senderName:
          type: string
        recipientName:
          type: string
        recipientEmail:
          type: string
        recipientPhone:
          type: string
        amount:
          type: number
        assetType:
          type: string
        status:
          type: string
          enum: [pending, delivered, redeemed, expired, cancelled]
        type:
          type: string
          enum: [digital, physical]
        message:
          type: string
        createdAt:
          type: string
        expiresAt:
          type: string
        redeemedAt:
          type: string
        campaignId:
          type: string
      required:
        - id
        - senderId
        - senderName
        - recipientName
        - recipientEmail
        - amount
        - assetType
        - status
        - type
        - createdAt
        - expiresAt
    TransactionListResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - transactions
        - total
        - page
        - limit
    Transaction:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        userName:
          type: string
        type:
          type: string
          enum: [purchase, gift, refund, settlement]
        amount:
          type: number
        status:
          type: string
          enum: [success, failed, pending, refunded]
        paymentMethod:
          type: string
        referenceId:
          type: string
        description:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        failureReason:
          type: string
      required:
        - id
        - userId
        - userName
        - type
        - amount
        - status
        - paymentMethod
        - createdAt
        - updatedAt
    CampaignListResponse:
      type: object
      properties:
        campaigns:
          type: array
          items:
            $ref: "#/components/schemas/Campaign"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - campaigns
        - total
        - page
        - limit
    Campaign:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [draft, scheduled, active, completed, cancelled]
        totalBudget:
          type: number
        spent:
          type: number
        totalRecipients:
          type: integer
        deliveredCount:
          type: integer
        redeemedCount:
          type: integer
        giftAmount:
          type: number
        assetType:
          type: string
        scheduledAt:
          type: string
        createdAt:
          type: string
        completedAt:
          type: string
        department:
          type: string
        customMessage:
          type: string
      required:
        - id
        - name
        - status
        - totalBudget
        - spent
        - totalRecipients
        - deliveredCount
        - redeemedCount
        - giftAmount
        - assetType
        - createdAt
    CreateCampaignRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        giftAmount:
          type: number
        assetType:
          type: string
        totalBudget:
          type: number
        scheduledAt:
          type: string
        department:
          type: string
        customMessage:
          type: string
      required:
        - name
        - giftAmount
        - assetType
        - totalBudget
...