# lib/api-zod/src/generated/types/giftOrder.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gfolio Admin & Corporate Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { GiftOrderStatus } from "./giftOrderStatus";
import type { GiftOrderType } from "./giftOrderType";

export interface GiftOrder {
  id: string;
  senderId: string;
  senderName: string;
  recipientName: string;
  recipientEmail: string;
  recipientPhone?: string;
  amount: number;
  assetType: string;
  status: GiftOrderStatus;
  type: GiftOrderType;
  message?: string;
  createdAt: string;
  expiresAt: string;
  redeemedAt?: string;
  campaignId?: string;
}